pyproject.toml
[project]
name = "agentic-devops-loop"
version = "0.2.0"
description = "Voice-to-Jira pipeline with agentic DevOps loop"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "faster-whisper>=1.0.1",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "python-socketio[asyncio]>=5.11.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
exclude = [
    ".claude/*",
    "venv",
    ".venv",
    "__pycache__",
    ".git",
    ".pytest_cache",
]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests for isolated components",
    "integration: Integration tests with external dependencies",
    "e2e: End-to-end workflow tests",
    "slow: Tests that take longer to run",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "tests/*",
    ".claude/*",
    "venv/*",
    ".venv/*",
    "scripts/*",
    "docs/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true
